Do not set query params if they are at default
